# nutito-api/openapi-schema.json
{"openapi":"3.0.0","paths":{"/api/v1":{"get":{"operationId":"AppController_home","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/users":{"get":{"operationId":"UserController_getUsers","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["auth"]}},"/api/v1/users/auth/signup":{"post":{"operationId":"UserController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupResponseDto"}}}}},"tags":["auth"]}},"/api/v1/users/auth/signin":{"post":{"operationId":"UserController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninResponseDto"}}}}},"tags":["auth"]}},"/api/v1/users/auth/mail/check":{"post":{"operationId":"UserController_verifyEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/v1/users/auth/check-email-existence/{email}":{"get":{"operationId":"UserController_checkIfEmailExists","parameters":[{"name":"email","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["auth"]}},"/api/v1/users/auth/check-phone-existence/{phone}":{"get":{"operationId":"UserController_checkIfPhoneExists","parameters":[{"name":"phone","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["auth"]}},"/api/v1/acccounts":{"get":{"operationId":"AccountController_getAccounts","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}}},"tags":["accounts"]},"post":{"operationId":"AccountController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}}},"tags":["accounts"]}},"/api/v1/acccounts/{code}":{"get":{"operationId":"AccountController_getAccountById","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}}},"tags":["accounts"]}},"/api/v1/acccounts/update/{id}":{"put":{"operationId":"AccountController_updateAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAccountDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}}},"tags":["accounts"]}},"/api/v1/acccounts/delete/{id}":{"delete":{"operationId":"AccountController_deleteAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["accounts"]}},"/api/v1/agents":{"get":{"operationId":"AgentController_getAgents","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Agent"}}}}}}}},"/api/v1/agents/{id}":{"get":{"operationId":"AgentController_getAgentById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Agent"}}}}}},"delete":{"operationId":"AgentController_deleteAgent","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"NUTITO SYSTEMS REST API","description":"This is the api for NUTITO SYSTEMS","version":"1.0","contact":{"name":"THE DEVCRAFT TEAM","url":"https://www.devcraft-team.com","email":"noreply@devcraft.com"}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"File":{"type":"object","properties":{"name":{"type":"string"},"location":{"type":"string"},"mimetype":{"type":"string"},"entity_name":{"type":"string"},"id":{"type":"number"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"creator_id":{"type":"number"},"updater_id":{"type":"number"},"slug":{"type":"string"}},"required":["name","location","mimetype","entity_name","id","created_at","updated_at","creator_id","updater_id","slug"]},"User":{"type":"object","properties":{"id":{"type":"number"},"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"ifu":{"type":"string"},"birth_date":{"format":"date-time","type":"string"},"birth_place":{"type":"string"},"address":{"type":"string"},"country":{"type":"string"},"city":{"type":"string"},"password":{"type":"string"},"user_type":{"type":"string"},"verification_token":{"type":"string"},"active":{"type":"boolean"},"profile_picture":{"$ref":"#/components/schemas/File"},"agents":{"type":"array","items":{"$ref":"#/components/schemas/Agent"}},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"creator_id":{"type":"number"},"updater_id":{"type":"number"},"slug":{"type":"string"}},"required":["id","firstname","lastname","email","phone","ifu","birth_date","birth_place","address","country","city","password","user_type","verification_token","active","agents","created_at","updated_at","creator_id","updater_id","slug"]},"CompanyCategory":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"companies":{"type":"array","items":{"$ref":"#/components/schemas/Company"}},"id":{"type":"number"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"creator_id":{"type":"number"},"updater_id":{"type":"number"},"slug":{"type":"string"}},"required":["name","description","id","created_at","updated_at","creator_id","updater_id","slug"]},"Licence":{"type":"object","properties":{"expiry_date":{"format":"date-time","type":"string"},"amount":{"type":"number"},"code":{"type":"string"},"payment_type":{"type":"string"},"transaction_id":{"type":"number"},"transaction_info":{"type":"string"},"company":{"$ref":"#/components/schemas/Company"},"id":{"type":"number"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"creator_id":{"type":"number"},"updater_id":{"type":"number"},"slug":{"type":"string"}},"required":["expiry_date","amount","code","payment_type","transaction_id","transaction_info","company","id","created_at","updated_at","creator_id","updater_id","slug"]},"Account":{"type":"object","properties":{"code":{"type":"string"},"amount":{"type":"number"},"amount_in":{"type":"number"},"amount_out":{"type":"number"},"id":{"type":"number"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"creator_id":{"type":"number"},"updater_id":{"type":"number"},"slug":{"type":"string"}},"required":["code","amount","amount_in","amount_out","id","created_at","updated_at","creator_id","updater_id","slug"]},"Workfield":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"companies":{"type":"array","items":{"$ref":"#/components/schemas/Company"}},"id":{"type":"number"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"creator_id":{"type":"number"},"updater_id":{"type":"number"},"slug":{"type":"string"}},"required":["name","description","id","created_at","updated_at","creator_id","updater_id","slug"]},"Company":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"city":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"ifu":{"type":"string"},"rccm":{"type":"string"},"agents":{"type":"array","items":{"$ref":"#/components/schemas/Agent"}},"category":{"$ref":"#/components/schemas/CompanyCategory"},"licences":{"type":"array","items":{"$ref":"#/components/schemas/Licence"}},"account":{"$ref":"#/components/schemas/Account"},"workfields":{"type":"array","items":{"$ref":"#/components/schemas/Workfield"}},"id":{"type":"number"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"creator_id":{"type":"number"},"updater_id":{"type":"number"},"slug":{"type":"string"}},"required":["name","description","city","email","phone","address","ifu","rccm","category","account","id","created_at","updated_at","creator_id","updater_id","slug"]},"AgentRole":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"agents":{"type":"array","items":{"$ref":"#/components/schemas/Agent"}},"id":{"type":"number"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"creator_id":{"type":"number"},"updater_id":{"type":"number"},"slug":{"type":"string"}},"required":["name","description","id","created_at","updated_at","creator_id","updater_id","slug"]},"Agent":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"},"company":{"$ref":"#/components/schemas/Company"},"role":{"$ref":"#/components/schemas/AgentRole"},"id":{"type":"number"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"creator_id":{"type":"number"},"updater_id":{"type":"number"},"slug":{"type":"string"}},"required":["user","company","role","id","created_at","updated_at","creator_id","updater_id","slug"]},"CreateUserDto":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"ifu":{"type":"string"},"birth_date":{"type":"string"},"birth_place":{"type":"string"},"address":{"type":"string"},"country":{"type":"string"},"password":{"type":"string"},"user_type":{"type":"string"}},"required":["phone","password"]},"SignupResponseDto":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"AuthenticateUserDto":{"type":"object","properties":{"phone":{"type":"string"},"password":{"type":"string"}},"required":["phone","password"]},"SigninResponseDto":{"type":"object","properties":{"status":{"type":"string"},"access_token":{"type":"string"},"user_data":{"type":"object"},"type":{"type":"string","default":"Bearer"}},"required":["status","access_token","user_data","type"]},"VerifyEmailDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/CreateUserDto"},"token":{"type":"string"}},"required":["user","token"]},"CreateAccountDto":{"type":"object","properties":{"code":{"type":"string"},"amount":{"type":"number"},"amount_in":{"type":"number"},"amount_out":{"type":"number"}},"required":["code","amount","amount_in","amount_out"]},"UpdateAccountDto":{"type":"object","properties":{"code":{"type":"string"},"amount":{"type":"number"},"amount_in":{"type":"number"},"amount_out":{"type":"number"}}}}}}