# nutito-api/src/accountancy/entities/agent-role.entity.ts
import { Column, Entity, OneToMany } from 'typeorm';
import { BaseEntity } from './base.entity';
import { Agent } from './agent.entity';

@Entity()
export class AgentRole extends BaseEntity {

    @Column('varchar')
    name: string;

    @Column('text')
    description: string;

    @OneToMany(type => Agent, agent => agent.role, { onDelete: "NO ACTION" })
    agents?: Agent[]

}